+++ src/redux/sagas/index.ts
@@ -1,31 +1,31 @@
 import { all, call } from 'redux-saga/effects';
-import watchAuthEmail from './AuthEmailSaga';
-import watchAuthCode from './AuthCodeSaga';
-import watchAuthUser from './AuthUserSaga';
-import watchSignUp from './SignUpSaga';
-import watchSignIn from './SignInSaga';
-import watchCategory from './CategorySaga';
-import watchSearch from './SearchSaga';
-import watchChangeNick from './ChangeNickSaga';
-import watchChangePassword from './ChangePasswordSaga';
-import watchRemoveUser from './RemoveUserSaga';
-import watchLibrary from './LibrarySaga';
-import watchLibraryAdd from './LibraryAddSaga';
-import watchLibraryModify from './LibraryModifySaga';
-import watchLibraryDelete from './LibraryDeleteSaga';
-import watchParagraph from './ParagraphSaga';
-import watchParagraphAdd from './ParagraphAddSaga';
-import watchParagraphDelete from './ParagraphDeleteSaga';
-import watchReview from './ReviewSaga';
-import watchReviewDetail from './ReviewDetailSaga';
-import watchNewReview from './NewReviewSaga';
-import watchUserReview from './UserReviewSaga';
-import watchUserReviewAdd from './UserReivewAddSaga';
-import watchUserReviewModify from './UserReviewModifySaga';
-import watchUserReviewDelete from './UserReviewDeleteSaga';
-import watchFavorite from './FavoriteSaga';
-import watchFavoriteAdd from './FavoriteAddSaga';
-import watchFavoriteCancel from './FavoriteCancelSaga';
+import watchAuthEmail from './authEmailSaga';
+import watchAuthCode from './authCodeSaga';
+import watchAuthUser from './authUserSaga';
+import watchSignUp from './signUpSaga';
+import watchSignIn from './signInSaga';
+import watchCategory from './categorySaga';
+import watchSearch from './searchSaga';
+import watchChangeNick from './changeNickSaga';
+import watchChangePassword from './changePasswordSaga';
+import watchRemoveUser from './removeUserSaga';
+import watchLibrary from './librarySaga';
+import watchLibraryAdd from './libraryAddSaga';
+import watchLibraryModify from './libraryModifySaga';
+import watchLibraryDelete from './libraryDeleteSaga';
+import watchParagraph from './paragraphSaga';
+import watchParagraphAdd from './paragraphAddSaga';
+import watchParagraphDelete from './paragraphDeleteSaga';
+import watchReview from './reviewSaga';
+import watchReviewDetail from './reviewDetailSaga';
+import watchNewReview from './newReviewSaga';
+import watchUserReview from './userReviewSaga';
+import watchUserReviewAdd from './userReivewAddSaga';
+import watchUserReviewModify from './userReviewModifySaga';
+import watchUserReviewDelete from './userReviewDeleteSaga';
+import watchFavorite from './favoriteSaga';
+import watchFavoriteAdd from './favoriteAddSaga';
+import watchFavoriteCancel from './favoriteCancelSaga';
 
 export default function* rootSaga() {
   yield all([
